simple_rag_project/requirements.txt
fastapi>=0.115.0
uvicorn>=0.30.0
python-dotenv>=1.0.0
sqlalchemy>=2.0.0
asyncpg>=0.29.0
aiosqlite>=0.20.0
redis>=5.0.0
qdrant-client>=1.10.0
sentence-transformers>=3.0.0
langchain-ollama>=0.2.0
langchain-openai>=0.2.0
langchain-google-genai>=2.0.0
langchain-text-splitters>=0.3.0
pydantic>=2.0.0
pydantic-settings>=2.0.0
pyjwt>=2.8.0
aiofiles>=24.0.0
pypdf2>=3.0.0
pdfplumber>=0.11.0
python-multipart>=0.0.9
tiktoken>=0.7.0
pillow>=10.0.0
email-validator>=2.0.0
